Create deleteme.html
The Tower of Hanoi puzzle is a classical problem that involves moving a set of different-sized disks from one rod to another, using an auxiliary rod, while adhering to specific rules. The primary challenge is to complete this task in the minimal number
of moves, which for n disks is (2^n) – 1. This project explores the implementation of the Tower of Hanoi solver using an Android application, aiming to provide not just a functional solution but also an educational tool to demonstrate the puzzle's algorithmic solution in action.